Docker + openshift deploy
