+++ .gitignore
@@ -1,2 +1,3 @@
-node_modules/
-settings.json
+**/node_modules
+**/.meteor/local
+app/settings.json
