+++ app/.meteor/versions
@@ -30,6 +30,7 @@ ecmascript-runtime-client@0.8.0
 ecmascript-runtime-server@0.7.1
 ejson@1.1.0
 email@1.2.3
+epfl:accounts-tequila@0.2.0
 erasaur:meteor-lodash@4.0.0
 es5-shim@4.8.0
 fetch@0.1.0
