+++ app/both/index.js
