+++ app/both/methods.js
