+++ app/client/ui/components/sites/Add.js
@@ -126,7 +126,8 @@ class Add extends React.Component {
           snowNumber: '',
           status:'requested',
           comment: '',
-          plannedClosingDate: ''
+          plannedClosingDate: '',
+          tags: []
         }
       }
