+++ app/package.json
@@ -20,7 +20,8 @@
     "react-router-dom": "^4.3.1",
     "react-select": "^2.4.0",
     "simpl-schema": "^1.5.3",
-    "yup": "^0.26.10"
+    "yup": "^0.26.10",
+    "p-wait-for": "*"
   },
   "meteor": {
     "mainModule": {
