+++ app/server/main.js
@@ -3,14 +3,43 @@ import { WebApp } from 'meteor/webapp';
 import { Sites } from '../both';
 import './publications';
 import getUnits from './units';
+import importVeritas from './import-data';
 
 // Define lang <html lang="fr" />
 WebApp.addHtmlAttributeHook(() => ({ lang: 'fr' }));
 
-let activeTequila = false;
+let activeTequila = true;
+/*
+Meteor.users.upsert(
+  { _id: '188', },
+  { 
+    $set: { 
+      username: 'charmier',
+      emails: ['gregory.charmier@epfl.ch'],
+     
+    }
+  }
+);
+
+// Add epfl-member by default
+if (!Roles.userIsInRole('188', ['admin', 'tags-editor', 'epfl-member'], Roles.GLOBAL_GROUP)) {
+  Roles.addUsersToRoles('188', 'admin', Roles.GLOBAL_GROUP);  
+}*/
   
 if (Meteor.isServer) {
+  
+  let nb = Sites.find({}).count();
 
+  console.log(process.env);
+  console.log(`INIT nb sites ${nb}`);
+  
+  if (nb == 0) {
+    console.log("Import data");
+    importVeritas();
+  } else {
+    console.log("Data already exist");
+  }
+  
   if (activeTequila) {
 
     Tequila.options.request = ['uniqueid', 'email'];
@@ -31,7 +60,7 @@ if (Meteor.isServer) {
       
       // Add epfl-member by default
       if (!Roles.userIsInRole(tequilaResponse.uniqueid, ['admin', 'tags-editor', 'epfl-member'], Roles.GLOBAL_GROUP)) {
-        Roles.addUsersToRoles(tequilaResponse.uniqueid, 'epfl-member', Roles.GLOBAL_GROUP);  
+        Roles.addUsersToRoles(tequilaResponse.uniqueid, 'admin', Roles.GLOBAL_GROUP);  
       }
 
       return tequilaResponse.uniqueid;
@@ -55,6 +84,15 @@ if (Meteor.isServer) {
   Api.addRoute('sites/:id', {authRequired: false}, {
     get: function () {
       return Sites.findOne(this.urlParams.id);
+    }
+  });
+
+  // TODO: à checker avec Luc car aujourd'hui title pas unique
+  Api.addRoute('sites-by-title/:title/tags', {authRequired: false}, {
+    get: function () {
+      
+      let site = Sites.findOne({title: this.urlParams.title});
+      return site.tags;
     }
   });
 
@@ -97,57 +135,5 @@ if (Meteor.isServer) {
   });
 
 
-importVeritas = () => {
-  const path = 'source-veritas.csv';
-  const file = Assets.getText(path);
-  Papa.parse(file, {
-    delimiter: ",",
-    header: true,
-    complete: function(results) {
-      
-      let data = JSON.parse(JSON.stringify(results.data));
-     
-      let index = 0;
-      data.forEach(site => {
-        index = index + 1;
-        console.log(index);
-
-        let langs;
-        if (site.langs == 'fr' || site.langs == 'en') {
-          langs = [site.langs];
-        } else {
-          langs = site.langs.split(',')
-        }
-
-        let siteDocument = {
-          url: site.wp_site_url,
-          tagline: site.wp_tagline,
-          title: site.wp_site_title,
-          openshiftEnv: site.openshift_env,
-          type: 'public',
-          category: null,
-          theme: site.theme,
-          faculty: site.theme_faculty,
-          languages: langs,
-          unitId: site.unit_id,
-          snowNumber: '',
-          status: 'created',
-          comment: '',
-          plannedClosingDate: null,
-          requestedDate: null,
-          createdDate: null,
-          archivedDate: null,
-          trashedDate: null,
-          tags: [],
-        }
-
-        if (!Sites.findOne({url: siteDocument.url})) {
-          Sites.insert(siteDocument);
-        }
-      });
-      console.log("Importation veritas finished");
-    }    
-  });
-}
 
 }
