+++ deploy-wp-veritas.yaml
@@ -1,0 +1,113 @@
+# 
+# Deployment for wp-veritas (without mongodb because we get it as a service)
+#
+---
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  annotations:
+    date: 2019-03-11 13:29
+    deployment.kubernetes.io/revision: '5'
+    kompose.cmd: kompose convert -f docker-compose.yml -o k8s-2.yml
+    kompose.version: 1.17.0 (a74acad)
+    version: 0.0.5
+  labels:
+    app: wp-veritas
+    author: gcharmier
+    team: idev-fsd
+    version: 0.0.5
+  name: wp-veritas
+  namespace: wwp-test
+spec:
+  replicas: 2
+  selector:
+    matchLabels:
+      app: wp-veritas
+  template:
+    metadata:
+      labels:
+        app: wp-veritas
+        author: gcharmier
+        team: idev-fsd
+        release: wp-veritas-mongo
+        jaime: lefromage
+    spec:
+      containers:
+        - env:
+            - name: MONGO_URL
+              valueFrom:
+                secretKeyRef:
+                  key: mongo-url
+                  name: wp-veritas
+            - name: PORT
+              value: '3000'
+            - name: ROOT_URL
+              value: 'https://wp-veritas.128.178.222.83.nip.io/'
+          image: 'epflidevelop/wp-veritas:0.1.10'
+          imagePullPolicy: IfNotPresent
+          name: pod-wp-veritas
+          ports:
+            - containerPort: 3000
+              protocol: TCP
+      dnsPolicy: ClusterFirst
+      restartPolicy: Always
+      schedulerName: default-scheduler
+      securityContext: {}
+      terminationGracePeriodSeconds: 30
+
+
+#
+# Sevice for wp-veritas
+#
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: service-wp-veritas
+  namespace: wwp-test
+  labels:
+    app: wp-veritas
+    author: gcharmier
+    team: idev-fsd
+  annotations:
+    authors: idev-fsd
+    version: 0.0.5
+spec:
+  ports:
+  - name: "80"
+    port: 80
+    protocol: TCP
+    targetPort: 3000
+  selector:
+    app: wp-veritas
+    jaime: lefromage
+  type: ClusterIP
+
+
+#
+# Route for wp-veritas
+#
+---
+apiVersion: route.openshift.io/v1
+kind: Route
+metadata:
+  name: route-wp-veritas
+  namespace: wwp-test
+  labels:
+    app: wp-veritas
+    author: gcharmier
+    team: idev-fsd
+  annotations:
+    authors: idev-fsd
+    version: 0.0.5
+spec:
+  host: wp-veritas.128.178.222.83.nip.io
+  port:
+    targetPort: "80"
+  tls:
+    termination: edge
+  to:
+    kind: Service
+    name: service-wp-veritas
+    weight: 100
+  wildcardPolicy: None
